Generative_AI/Langchain SearchBot With Tools and Agents/requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
ipykernel
langchain-groq
arxiv
wikipedia
duckdb
langchainhub
duckduckgo-search
